+++ VoiceOver Designer/SingleDocument/ProjectStateController+Toolbar.swift
@@ -12,7 +12,6 @@ extension ProjectStateController: NSToolbarDelegate {
         willBeInsertedIntoToolbar flag: Bool
     ) -> NSToolbarItem? {
         switch itemIdentifier {
-//        case .voiceControlLabel: return labelItem()
         case .documentsButtonLabel: return documentsItem()
         case .trailingSidebar: return trailingSideBarItem()
         case .shareDocument: return shareDocumentItem()
@@ -77,26 +76,10 @@ extension ProjectStateController: NSToolbarDelegate {
 }
 
 extension ProjectStateController {
-//    private func labelItem() -> NSToolbarItem {
-//        let item = NSToolbarItem(itemIdentifier: .voiceControlLabel)
-//        item.label = NSLocalizedString("Labels", comment: "")
-//        item.enableLabels()
-//        item.target = self
-//        item.action = #selector(showLabels(sender:))
-//        item.isBordered = true
-//        return item
-//    }
-    
     private func documentsItem() -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .documentsButtonLabel)
-        item.label = NSLocalizedString("Recent", comment: "")
+        let item = NSToolbarItem.makeDocumentsItem()
         item.target = self
         item.action = #selector(showRecentDidPressed(sender:))
-        item.isBordered = true
-        item.image = NSImage(systemSymbolName: "rectangle.grid.3x2",
-                             accessibilityDescription: "Show recent documents")!
-        item.toolTip = NSLocalizedString("Go to my documents", comment: "")
-        item.isNavigational = true
         return item
     }
     
@@ -109,26 +92,16 @@ extension ProjectStateController {
     
     @available(macOS, obsoleted: 14.0)
     private func trailingSideBarItem() -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .trailingSidebar)
-        item.label = NSLocalizedString("Inspector", comment: "")
+        let item = NSToolbarItem.trailingSideBarItem()
         item.target = self
         item.action = #selector(trailingSideBarTapped(sender:))
-        item.isBordered = true
-        item.image = NSImage(systemSymbolName: "sidebar.trailing",
-                             accessibilityDescription: "Close inspector sidebar")!
-        item.toolTip = NSLocalizedString("Close inspector sidebar", comment: "")
         return item
     }
 
     private func presentationSideBarItem() -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .presentation)
-        item.label = NSLocalizedString("Presentation", comment: "")
+        let item = NSToolbarItem.presentationSideBarItem()
         item.target = self
         item.action = #selector(presentationModeTapped(sender:))
-        item.isBordered = true
-        item.image = NSImage(systemSymbolName: "play.fill",
-                             accessibilityDescription: "Open presentation mode")!
-        item.toolTip = NSLocalizedString("Open presentation mode", comment: "")
         item.menuFormRepresentation = playMenuItem
         return item
     }
@@ -229,32 +202,8 @@ extension NSToolbarItem {
     }
 }
 
-extension NSToolbarItem.Identifier {
-//    static let voiceControlLabel = NSToolbarItem.Identifier(rawValue: "VoiceControlLabel")
-    static let documentsButtonLabel = NSToolbarItem.Identifier(rawValue: "DocumentsButtonLabel")
-    static let trailingSidebar = NSToolbarItem.Identifier(rawValue: "TrailingSidebar")
-    static let presentation = NSToolbarItem.Identifier(rawValue: "Presentation")
-    static let editor = NSToolbarItem.Identifier(rawValue: "Editor")
-    static let leadingSidebar = NSToolbarItem.Identifier(rawValue: "LeadingSidebar")
-    static let itemListTrackingSeparator = NSToolbarItem.Identifier("ItemListTrackingSeparator")
-    static let shareDocument = NSToolbarItem.Identifier("ShareDocument")
-}
-
 extension VODesignDocument: NSSharingServicePickerToolbarItemDelegate {
     public func items(for pickerToolbarItem: NSSharingServicePickerToolbarItem) -> [Any] {
         [fileURL as Any]
-    }
-}
-
-private extension NSToolbar {
-
-    @discardableResult
-    func removeItem(identifier: NSToolbarItem.Identifier) -> Int? {
-        if let index = items
-            .firstIndex(where: { $0.itemIdentifier == identifier }) {
-            removeItem(at: index)
-            return index
-        }
-        return nil
     }
 }
